# Cloud Computing/src/config/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Backend App Sobat Tani (Auth User, Bookmark, Status, Detail Plant)",
        "version": "1.0.0",
        "title": "Dokumentasi API Sobat Tani",
        "contact": {}
    },
...
    "paths": {
...
        "/auth/user": {
            "get": {
                "tags": [
                    "User Auth"
                ],
                "summary": "Detail User",
                "description": "# Get User Details\n\nRetrieves detailed information of the authenticated user\n\n### Request\n\n- Method: `GET`\n    \n- URL: `/auth/user`\n    \n- Headers:\n    \n    - `Authorization`: `Bearer`",
                "operationId": "detailUser",
                "produces": [
                    "application/json"
                ],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Success",
                        "headers": {
                            "Alt-Svc": {
                                "type": "string"
                            },
                            "Content-Length": {
                                "type": "string"
                            },
                            "Date": {
                                "type": "string"
                            },
                            "Server": {
                                "type": "string"
                            },
                            "X-Cloud-Trace-Context": {
                                "type": "string"
                            },
                            "access-control-allow-origin": {
                                "type": "string"
                            },
                            "etag": {
                                "type": "string"
                            },
                            "x-powered-by": {
                                "type": "string"
                            }
                        },
                        "schema": {
                            "$ref": "#/definitions/inline_response_200_3"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "User Auth"
                ],
                "summary": "updateUser",
                "description": "# Update User Information\n\nUpdates the profile information of an authenticated user  \n\nRequest\n\n- Method: `PUT`\n    \n- URL: `/auth/user`\n    \n- Headers:\n    \n    - `Authorization`: `Bearer`",
                "operationId": "updateuser",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/body_3"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "headers": {
                            "Alt-Svc": {
                                "type": "string"
                            },
                            "Content-Length": {
                                "type": "string"
                            },
                            "Date": {
                                "type": "string"
                            },
                            "Server": {
                                "type": "string"
                            },
                            "X-Cloud-Trace-Context": {
                                "type": "string"
                            },
                            "access-control-allow-origin": {
                                "type": "string"
                            },
                            "etag": {
                                "type": "string"
                            },
                            "x-powered-by": {
                                "type": "string"
                            }
                        },
                        "schema": {
                            "$ref": "#/definitions/inline_response_200_4"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "body_3": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": " C241-PS007new"
                },
                "password": {
                    "type": "string",
                    "example": "newsecurepassword01234"
                }
            }
        },
...
        "inline_response_200_3_data": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "sobattani1@gmail.com"
                },
                "name": {
                    "type": "string",
                    "example": " C241-PS007new"
                },
                "uid": {
                    "type": "string",
                    "example": "user-id"
                }
            }
        },
...
        "inline_response_200_3": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/inline_response_200_3_data"
                },
                "message": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "inline_response_200_4": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/inline_response_200_3_data"
                },
                "message": {
                    "type": "string",
                    "example": "Success"
                }
            }
        }
    }
}